adds metric table / class

+++ src/main/java/com/appdetex/entity/Metric.java
@@ -33,7 +33,7 @@ public class Metric {
     private int min;
 
     @Column(name="avg")
-    private int avg;
+    private double avg;
 
     @Column(name="sum")
     private int sum;
